Avoid using core module 'mediawiki.special.search'
The core module 'mediawiki.special.search' contains more functions
which are not needed here and which causes an error message in the
JavaScript console. (T211384)

This change copies the used part of module 'mediawiki.special.search'
to the local module 'ext.ReplaceText'.

The CSS file for the style module 'ext.ReplaceTextStyles' is renamed
from ext.ReplaceText.css to ext.ReplaceTextStyles.css to make
ext.ReplaceText.css free for the CSS part of module 'ext.ReplaceText'.

Bug: T211384
Change-Id: I41225ccdf8a95a7c501fb6eea99abbd08353f4ea