Add support for slots other than 'main'
This commit adds support for replacing text in multiple content slots.
The names of all the slots that contain the target string are shown on
the special page, and the user can select or deselect which slots they
want to edit. Edits to different slots for the same page are bundled
into a multi-content revision.

If a replace is performed through the 'replaceAll.php' maintenance
script, all slots are taken into account.

Change-Id: Ic4f36fa76e1eaeac8200c60f196c53c2ed78fa45