Fix default checking of main namespace
The old code "worked", but apparently just by accident, in that in_array( 0, [ null ] ) returns true. Somehow that code lasted in MediaWiki's Special:Search for 10 years or so - I guess no one questioned it, because it worked. But in at least some cases (with PHP 8, maybe) this doesn't work, so it's good to fix it.

Change-Id: I38dabec83ea350f65edf77ef160c7903c3894917